AI-scoring-model/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0

# Training dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
